% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_port_ids.R
\name{get_port_ids}
\alias{get_port_ids}
\title{Get the latest list of ports id's}
\usage{
get_port_ids()
}
\description{
Queries potential port ID's from the Portuguese National Hydrographic Institute and returns a data.frame with valid ones
}
