% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tides.R
\name{get_tides}
\alias{get_tides}
\title{Get tide table for a port and day}
\usage{
get_tides(
  port_id = 19,
  start_date = Sys.Date(),
  end_date = NULL,
  day_range = 1,
  include_moons = FALSE,
  silent = FALSE
)
}
\arguments{
\item{port_id}{The id code for the desired port. Use `port_list()` to see a list of IDs. Defaults to 19, which is Faro-Olhão.}

\item{start_date}{The starting date for the wanted tides. Format should be yyyy-mm-dd or yyyy/mm/dd. Accepts class character, Date and POSIXct. Defaults to current date.}

\item{end_date}{The end date for the tidal table. Same format as start_date}

\item{day_range}{OPTIONAL You can skip the end date and just say how many days you are interested in. Defaults to 1 which only provides tides for `start_date`}

\item{include_moons}{Should lunar events be kept in the table? Defaults to FALSE.}

\item{silent}{Should messages be suppressed? Defaults to FALSE (display messages).}
}
\description{
Returns a data.frame with all tidal, for a time period with a specified start date and duration. The returned times are always in the local GMT time for the port (as specified in by the Portuguese National Hydrographic Institute).
}
\examples{
Retrieve the information for the Faro - Olhao port, for 7 days, starting at March 5th of 2020
tides <- get_tides(port_id = 19, date = "2020-03-05", day_range = 7)

}
