% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_port_ids.R
\docType{data}
\name{homepage}
\alias{homepage}
\title{Get the latest list of ports id's}
\format{An object of class \code{character} of length 1.}
\usage{
homepage
}
\description{
Queries potential port ID's from the Portuguese National Hydrographic Institute and returns a data.frame with valid ones
}
\keyword{datasets}
