% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tides.R
\name{plot_tides}
\alias{plot_tides}
\title{Plot a calendar of the tidal events for each day}
\usage{
plot_tides(tides)
}
\arguments{
\item{tides}{The tides table, obtained from get_tides}
}
\description{
If you pass it a tides table (from get_tides), a ggplot of the high and low tides will be created. I suggest adding vertical lines (geom_vline) to make visualization easier
}
\examples{
plot_tides(tides_table)


}
