% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_list.R
\name{port_list}
\alias{port_list}
\title{Show list of port ids}
\usage{
port_list()
}
\description{
Returns a data.frame with the port ID's used in get_tides
Based on data scrapped on 2020/03/13
If port_id is not working, use get_port_ids to generate an updated list
}
\examples{
port_list()

}
