% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_table.R
\name{tide_tabe}
\alias{tide_tabe}
\title{Get tide table for a port and day}
\usage{
tide_tabe(port_id = NULL, date = NULL, day_range = NULL)
}
\arguments{
\item{port_id}{The id code for the desired port (use valid_ports to see a list)}

\item{date}{The starting date for the wanted tides. Format should be yyyy-mm-dd or yyyy/mm/dd}

\item{day_range}{The number of days after date for which to retrieve information}
}
\description{
This function retrieves the tides for the desired dates and time range
}
