% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide_table.R
\name{tide_table}
\alias{tide_table}
\title{Get tide table for a port and day}
\usage{
tide_table(port_id = 19, date = NULL, day_range = NULL, include_moons = FALSE)
}
\arguments{
\item{port_id}{The id code for the desired port (use valid_ports to see a list)}

\item{date}{The starting date for the wanted tides. Format should be yyyy-mm-dd or yyyy/mm/dd}

\item{day_range}{The number of days after date for which to retrieve information}

\item{include_moons}{Should lunar events be kept in the table? TRUE or FALSE}
}
\description{
This function retrieves the tides for the desired dates and time range.
By default, data for the Faro - Olhão port, for the current day, is retrieved.
}
