% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_ports.R
\name{valid_ports}
\alias{valid_ports}
\title{Show valid ports}
\usage{
valid_ports()
}
\description{
Returns a data.frame with the port ID's required to retrieve tide_tables
This is a brute force method, querying the server for 1000 possible id numbers. 
While 1000 queries are not a crazy ammount, I advise you to minimize the use of this function.
}
\examples{
valid_ports()
}
