% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tides.R
\name{get_tides}
\alias{get_tides}
\title{Get tide table for a port and day}
\usage{
get_tides(port_id = 19, date = NULL, day_range = NULL, include_moons = FALSE)
}
\arguments{
\item{port_id}{The id code for the desired port. Use valid_ports to see a list of ids.}

\item{date}{The starting date for the wanted tides. Format should be yyyy-mm-dd or yyyy/mm/dd}

\item{day_range}{The number of days after date for which to retrieve information}

\item{include_moons}{Should lunar events be kept in the table? TRUE or FALSE}
}
\description{
This function retrieves the tides for the desired dates and time range. By default, data for the Faro - Olhao port, for the current day, is retrieved.
}
\examples{
Retrieve the information for the Faro - Olhao port, for 7 days starting at March 5th of 2020
tide_table(port_id = 19, date = "2020-03-05", day_rage = 7)

}
