% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_tides.R
\name{interpolate_tides}
\alias{interpolate_tides}
\title{Takes a vector of date times and returns the expected tide for a port}
\usage{
interpolate_tides(date_time = NULL, port_id = NULL)
}
\arguments{
\item{port_id}{The id code for the desired port (use valid_ports to see a list)}

\item{date}{The starting date for the wanted tides. Format should be yyyy-mm-dd or yyyy/mm/dd}

\item{day_range}{The number of days after date for which to retrieve information}
}
\description{
This function retrieves the tides for the desired dates and time range
}
