% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_tides.R
\name{interpolate_tides}
\alias{interpolate_tides}
\title{Calculate estimated tide heights}
\usage{
interpolate_tides(date_times = NULL, port_id = 19)
}
\arguments{
\item{date_times}{A character vector with the format yyyy-mm-dd hh:mm:ss OR a POSIXct vector}

\item{port_id}{The id code for the desired port (use port_list to see a list, Faro-Olhão is the default)}
}
\description{
This function takes a series of time-points and a port ID.
It returns the expected tide height at those time-points, for that port
based on the method provided by the Portuguese National Hydrographic Institute
}
\examples{
Retrieve the information for the Faro - Olhao port, for the 7 days after March 5th of 2020
sampling_times <- c("2020-03-13 15:15:00", "2020-03-09 16:15:00")
interpolate_tides(date_times = sampling_times, port_id = 19)
}
